+++ packages/asc-ui/src/index.ts
@@ -75,7 +75,13 @@ import {
 import Modal from './components/Modal'
 import Paragraph, { ParagraphStyle } from './components/Paragraph'
 import Portal from './components/Portal'
-import Radio, { RadioGroup } from './components/Radio'
+import Quote, { QuoteStyle } from './components/Quote'
+import Radio, {
+  RadioGroup,
+  RadioStyle,
+  RadioWrapperStyle,
+  RadioCircleStyle,
+} from './components/Radio'
 import SearchBar, { SearchBarStyles } from './components/SearchBar'
 import SearchBarToggle, {
   SearchBarToggleStyles,
@@ -135,6 +141,7 @@ export {
 export { constants }
 export { perceivedLoading }
 // Components
+export { Quote, QuoteStyle }
 export {
   Accordion,
   AccordionWrapper,
@@ -259,6 +266,9 @@ export const styles = {
   CheckboxStyle,
   FilterBoxStyle,
   FilterOptionStyle,
+  RadioStyle,
+  RadioWrapperStyle,
+  RadioCircleStyle,
 }
 
 export const context = {